Refactor UserPlantForm and UserPlant model to remove unused import and commented code
